requirements.txt
llama-cpp-python==0.2.90
pymysql==1.1.1
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
